login using google auth
OK, so this allows users to login using google sign in button. It wont work if you try to run the program just like that. If you want it to work immediately, get rid of the MainActivity class, and replace all MainActivity words with LoginActivity. Later, we shall link loginactivity class to mainactivity.